% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KayaIdentity.R
\name{kaya_eq}
\alias{kaya_eq}
\title{Equation for Kaya Identity}
\usage{
kaya_eq(pop, gdp, enInt, carbInt, outputType = "CO2")
}
\arguments{
\item{pop}{population size (in millions), input as non-negative numeric}

\item{gdp}{GDP per person (in 1000$/person), input as non-negative numeric}

\item{enInt}{energy intensity (in Gigajoule/$1000GDP), input as non-negative numeric}

\item{carbInt}{carbon intensity (in tonnes CO2/Gigajoule), input as non-negative numeric}

\item{outputType}{output type either as CO2 or C emission}
}
\value{
numerical value representing the yearly CO2 emissions in million tonnes
}
\description{
The kaya identity expresses yearly CO2 emissions as a product of four factors
}
\examples{
# the yearly CO2 emissions of Germany
kaya_eq(pop = 82.4, gdp = 44, enInt = 5, carbInt = 0.05)
# the yearly C emissions of Germany
kaya_eq(pop = 82.4, gdp = 44, enInt = 5, carbInt = 0.05, outputType = "C")
}
